requirements.txt
Flask
Flask-Cors
protobuf
pycryptodome
httpx
cachetools
pytz
